class-5.10-5.11
৫.১০ - অথর পেজ - এবং অথর টেমপ্লেট হায়ারার্কি
৫.১১ - অ্যাটাচমেন্টস প্লাগইনের সাথে পরিচয় এবং এর সাহায্যে চমৎকার একটা স্লাইডার তৈরী করা

+++ app/public/wp-content/themes/alpha/single.php
@@ -31,7 +31,7 @@ if (!is_active_sidebar("sidebar-1")) {
                                     <?php the_title();?>
                                 </h2>
                                 <p class="<?php echo $alpha_text_class?>">
-                                    <strong><?php the_author(); ?></strong><br/>
+                                    <strong><?php the_author_posts_link(); ?></strong><br/>
                                     <?php echo get_the_date();?>
                                 </p>
                             </div>
@@ -39,26 +39,43 @@ if (!is_active_sidebar("sidebar-1")) {
                         <div class="row">
                             
                             <div class="col-md-12">
-                                <p>
+                                <div class="slider">
                                     <?php
-                                        //for post thumbnail/image
-                                        if(has_post_thumbnail()) {
-                                            //$thumbnail_url = get_the_post_thumbnail_url( null, "large" );
-                                            // echo "<a href="'.$thumbnail_url.'" data-featherlight="image">";
-                                            echo '<a class="phpup" href="#" data-featherlight="image">';
-                                            the_post_thumbnail("large", array("class"=>"img-fluid"));
-                                            echo "</a>";
+                                    if ( class_exists( 'Attachments' ) ) {
+                                        $attachments = new Attachments( 'slider' );
+                                        if ( $attachments->exist() ) {
+                                            while ($attachment = $attachments->get() ) {
+                                                ?>
+                                                    <div>
+                                                        <?php echo $attachments->image( 'large' );?>
+                                                    </div>
+                                                <?php
+                                            }
+                                        }
+                                    }
+                                    ?>
+                                </div>
+                                <div>
+                                    <?php
+                                        if ( !class_exists( 'Attachments' ) ) {
+                                            // for post thumbnail/image
+                                            if(has_post_thumbnail()) {
+                                                //$thumbnail_url = get_the_post_thumbnail_url( null, "large" );
+                                                // echo "<a href="'.$thumbnail_url.'" data-featherlight="image">";
+                                                echo '<a class="phpup" href="#" data-featherlight="image">';
+                                                the_post_thumbnail("large", array("class"=>"img-fluid"));
+                                                echo "</a>";
+                                            }
                                         }
 
                                         the_content();
-
                                         wp_link_pages();
 
                                         /*/next_post_link();
                                         echo "<br/>";
                                         previous_post_link();*/
                                     ?>
-                                </p>
+                                </div>
                                 
                             </div>
